// src/app/model/policyMaster.ts
export class policyMaster {
  id: number = 0;
  pid: string = '';
  propertyType: string = '';
  consumer_type: string = '';
  assured_sum: string = '';
  tenure: string = '';
  businessValue: number = 0;
  propertyValue: number = 0;
  base_location: string = '';
  type: string = '';
  public constructor() {}
}